requirements.txt
# Core
pandas==3.0.2
requests==2.33.1
python-dotenv==1.2.2

# AWS / S3
boto3
s3fs==2026.3.0
aiobotocore==3.4.0
botocore==1.42.84

# Airflow
apache-airflow==3.1.8
apache-airflow-providers-standard==1.12.2